Added disclaimer about "trial kit" toy silicone quantity
I felt it was important to note that although the shopping list indicates "trial size" as the amount of toy silicone to get, that it really depends on the size of a particular design.

IME a 28-oz trial kit is enough for 1-2 copies of moderately sized toys. I'm still hesitant to say even that though, because people have differing ideas about toy size baselines..