Change marbling mix image to black and gold
The previous image was of the nebula colors, which was potentially confusing since the "pour" section earlier in the document depicted those colors being poured -- a backwards chronology. The existing picture was also on the high end of acceptable mixing.

The new image showing black and gold should be a little clearer, I think.