Revised "toy silicones" section
I bumped Ecoflex 00-50 to the default recommendation instead of DS10M. This is both because "medium" firmness seems like a better thing to recommend people start with, as well as 00-50 having a much lower viscosity which makes it easier to work with.

I also added a note recommending people get a set of silicone sample disks to use as reference.